Authentication fixes, error proofing in userService

+++ organizer/src/main/java/org/backend/organizer/Service/UserService.java
@@ -1,5 +1,6 @@
 package org.backend.organizer.Service;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.backend.organizer.Model.User;
 import org.backend.organizer.Model.UserPrincipal;
 import org.backend.organizer.Repository.UserRepository;
@@ -34,7 +35,7 @@ public class UserService {
 
     public User getUserById(Long id) {
         if(id == null) throw new NullPointerException();
-        return repository.getUserById(id); // does not work with getReferenceById
+        return repository.findById(id).orElseThrow(EntityNotFoundException::new);
     }
 
     public User register(User user) {
@@ -55,7 +56,7 @@ public class UserService {
             ResponseCookie jwtCookie = jwtService.generateJwtCookie(userPrincipal);
             return jwtCookie.toString();
         } else {
-            return "fail";
+            return "fail"; // request is 401 unauthorized anyway
         }
     }
 
@@ -68,15 +69,18 @@ public class UserService {
     }
 
     public User updateUser(User userUpdates) {
-        User user = repository.getUserById(userUpdates.getId());
+        User user = repository.findById(userUpdates.getId()).orElseThrow(EntityNotFoundException::new);
+        //TODO use mapper to ignore nulls
         if (userUpdates.getEmail() != null) user.setEmail(userUpdates.getEmail());
         if (userUpdates.getName() != null) user.setName(userUpdates.getName());
         if (userUpdates.getUsername() != null) user.setUsername(userUpdates.getUsername());
         if (userUpdates.getPassword() != null) user.setPassword(encoder.encode(userUpdates.getPassword()));
+        if (userUpdates.getRole() != null) user.setRole(userUpdates.getRole());
         return repository.save(user);
     }
 
     public void deleteUser(Long id) {
+        if(!repository.existsById(id)) throw new EntityNotFoundException();
         repository.deleteById(id);
     }
 }
